requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
mutagen==1.47.0
requests==2.31.0
gunicorn==21.2.0
boto3
python-dotenv